pyproject.toml
[tool.poetry]
name = "learnwiki_backend"
version = "0.1.0"
description = ""
authors = ["TeaDove <peter.ibragimov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
loguru = "^0.5.3"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
django-extensions = "^3.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
